Tiny changes for running from no state.
Note you have to make the output directory.

Guard against an empty database.

+++ mastodon/build.py
@@ -52,7 +52,8 @@ maxsql = """
 maxget = cur.execute(maxsql)
 maxsnap = maxget.fetchone()["maxsnap"]
 
-cleansql = "DELETE FROM links WHERE snapshot != " + str(maxsnap) +  ";"
+if maxsnap is not None:
+	cleansql = "DELETE FROM links WHERE snapshot != " + str(maxsnap) +  ";"
 
 cur.execute(cleansql)
 con.commit()
